[DLTv2][UPGRADE] Adapt ISO C/C++ for DLTv2
- Enable fail fast -Wfatal-error
- Fix abuse casting for data
- Fix wrong function usages
- Fix critical potential use-after-free of msgv2.headerbufferv2
- Rebase and solve merge conflicts
- Adapt casting in regression test binaries
- Change dlt_set_id_v2() API to perform fixed-length ID setup

+++ src/examples/dlt-example-user-func-v2.c
@@ -287,7 +287,7 @@ int dlt_user_injection_callback(uint32_t service_id, void *data, uint32_t length
     printf("Injection %d, Length=%d \n", service_id, length);
 
     if (length > 0) {
-        dlt_print_mixed_string(text, 1024, data, length, 0);
+        dlt_print_mixed_string(text, 1024, data, (int)length, 0);
         printf("%s \n", text);
     }
